+++ tests/YandexDirect/Event/FailCallEventTest.php
@@ -12,17 +12,28 @@ class FailCallEventTest extends TestCase
     public function testConstructorAndGetters()
     {
         $user = $this->getUserMock();
+        $client = $this->getSoapClientMock();
         $methodName = 'Foo';
         $methodParams = array('FooParam');
         $requestId = 'request-id';
         $exception = new \Exception();
 
-        $event = new FailCallEvent($methodName, $methodParams, $user, $requestId, $exception);
+        $client->expects($this->any())
+            ->method('getLastRequest')
+            ->willReturn('request content');
+
+        $client->expects($this->any())
+            ->method('getLastResponse')
+            ->willReturn('response content');
+
+        $event = new FailCallEvent($methodName, $methodParams, $user, $requestId, $client, $exception);
 
         $this->assertSame($user, $event->getUser());
         $this->assertSame($methodName, $event->getMethodName());
         $this->assertSame($methodParams, $event->getMethodParams());
         $this->assertSame($exception, $event->getException());
         $this->assertSame($requestId, $event->getRequestId());
+        $this->assertSame('request content', $event->getRequest());
+        $this->assertSame('response content', $event->getResponse());
     }
 }