+++ tests/YandexDirect/Event/PostCallEventTest.php
@@ -12,17 +12,28 @@ class PostCallEventTest extends TestCase
     public function testConstructorAndGetters()
     {
         $user = $this->getUserMock();
+        $client = $this->getSoapClientMock();
         $methodName = 'Foo';
-        $responce = 'Any responce';
+        $result = 'Any responce';
         $params = array('foo' => 'bar');
         $requestId = 'request-id';
 
-        $event = new PostCallEvent($methodName, $params, $user, $requestId, $responce);
+        $client->expects($this->any())
+            ->method('getLastRequest')
+            ->willReturn('request content');
+
+        $client->expects($this->any())
+            ->method('getLastResponse')
+            ->willReturn('response content');
+
+        $event = new PostCallEvent($methodName, $params, $user, $requestId, $client, $result);
 
         $this->assertSame($user, $event->getUser());
         $this->assertSame($methodName, $event->getMethodName());
-        $this->assertSame($responce, $event->getResponse());
+        $this->assertSame($result, $event->getResult());
         $this->assertSame($params, $event->getMethodParams());
         $this->assertSame($requestId, $event->getRequestId());
+        $this->assertSame('request content', $event->getRequest());
+        $this->assertSame('response content', $event->getResponse());
     }
 }